Add TLS support
This adds support for processing HTTPS requests on a separate port from
the usual HTTP requests. It does so by adding options for specifying the
path to a TLS key and a TLS certificate.

This can be useful for testing worker scripts that do things like custom
redirects depending on whether the user is making an encrypted request
or not.

Note: This PR does not yet set special `request.cf` attributes such as
`tlsVersion` and `tlsCipher` as described here:
https://developers.cloudflare.com/workers/reference/request-attributes/